requirements.txt
--index-url https://pypi.python.org/simple/
-e ./vendor/geonode
-e ./vendor/gsconfig
-e ./vendor/django-osgeo-importer
-e ./vendor/arcrest/src
-e ./vendor/django-ssl-pki
-e ./vendor/geonode-elasticsearch
git+https://github.com/MapStory/composer.git@master#composer
git+https://github.com/boundlessgeo/sphinx@bex#egg=Sphinx
git+https://github.com/venicegeo/nearsight@master#egg=nearsight
django-jsonfield==1.0.1
django-jsonfield-compat==0.4.4
numpy==1.13.3
django-geonode-client==1.0.9
dj-database-url==0.4.2
celery[redis]==3.1.18
django-storages==1.6.5
boto==2.48.0
waitress==0.9.0
whitenoise==3.3.1
django-cors-headers==1.3.1
django-classification-banner==0.1.5
django-solo==1.1.2
django-colorfield==0.1.14
psycopg2==2.7.3.2
python-ldap==2.5.2
django-auth-ldap==1.3.0
GDAL==2.3.1
supervisor==3.3.3
supervisor-stdout==0.1.1
python-resize-image==1.1.11
django-flat-theme==1.1.4
flake8==3.5.0
social-auth-app-django==1.2.0
mock==2.0.0
Django==1.8.18
six==1.10.0
cryptography==2.3
urllib3==1.23
regex==2018.11.03
django-autocomplete-light==3.3.2
git+https://github.com/jpic/xmodelform.git@master#egg=xmodelform

# For pki app
# should be with [secure] for pyOpenSSL, cryptography, etc. for modern SSL
# urllib3[secure]==1.22
# minimum of requests 2.12 needed, though not sure if it is urllib3-bug-free
requests==2.20.0
# more robust SSL in urllib3 with pyOpenSSL
# pyOpenSSL==17.5.0
# PKI component validations and db field encryption
# cryptography==2.1.4
# for user-ordering of wildcard hostname:port mappings
django-ordered-model==1.4.3
